requirements.txt
watchdog>=3.0.0
aiofiles==24.1.0
ollama==0.4.8
Pillow==11.2.1
python-telegram-bot==22.0
websocket_client==1.8.0
